unit & integration tests dor delete feature

+++ src/main/java/com/tonasolution/appointnear/business/AppointmentServiceImp.java
@@ -46,6 +46,7 @@ public class AppointmentServiceImp implements AppointmentService {
 
 	@Override
 	public void delete(Long id) {
+		if (id == null) throw new IllegalArgumentException("Bad attribute");
 		this.appointmentDao.deleteById(id);
 		
 	}
